JSON parser decimal type changed from double to float

+++ lite3dpp/include/lite3dpp/json/JSON.h
@@ -27,9 +27,9 @@
 
 // Win32 incompatibilities
 #if defined(WIN32) && !defined(__GNUC__)
-	#define wcsncasecmp _wcsnicmp
-	static inline bool isnan(double x) { return x != x; }
-	static inline bool isinf(double x) { return !isnan(x) && isnan(x - x); }
+    #define wcsncasecmp _wcsnicmp
+    static inline bool isnan(float x) { return x != x; }
+    static inline bool isinf(float x) { return !isnan(x) && isnan(x - x); }
 #endif
 
 #include <vector>
@@ -38,53 +38,53 @@
 
 // Linux compile fix - from quaker66
 #ifdef __GNUC__
-	#include <cstring>
-	#include <cstdlib>
+    #include <cstring>
+    #include <cstdlib>
 #endif
 
 #include <lite3dpp/lite3dpp_manageable.h>
     
 // Mac compile fixes - from quaker66, Lion fix by dabrahams
 #if defined(__APPLE__) && __DARWIN_C_LEVEL < 200809L || (defined(WIN32) && defined(__GNUC__)) || defined(ANDROID)
-	#include <wctype.h>
-	#include <wchar.h>
-	
-	static inline int wcsncasecmp(const wchar_t *s1, const wchar_t *s2, size_t n)
-	{
-		int lc1  = 0;
-		int lc2  = 0;
+    #include <wctype.h>
+    #include <wchar.h>
+    
+    static inline int wcsncasecmp(const wchar_t *s1, const wchar_t *s2, size_t n)
+    {
+        int lc1  = 0;
+        int lc2  = 0;
 
-		while (n--)
-		{
-			lc1 = towlower (*s1);
-			lc2 = towlower (*s2);
+        while (n--)
+        {
+            lc1 = towlower (*s1);
+            lc2 = towlower (*s2);
 
-			if (lc1 != lc2)
-				return (lc1 - lc2);
+            if (lc1 != lc2)
+                return (lc1 - lc2);
 
-			if (!lc1)
-				return 0;
+            if (!lc1)
+                return 0;
 
-			++s1;
-			++s2;
-		}
+            ++s1;
+            ++s2;
+        }
 
-		return 0;
-	}
+        return 0;
+    }
 #endif
 
 // Simple function to check a string 's' has at least 'n' characters
 static inline bool simplejson_wcsnlen(const wchar_t *s, size_t n) {
-	if (s == 0)
-		return false;
+    if (s == 0)
+        return false;
 
-	const wchar_t *save = s;
-	while (n-- > 0)
-	{
-		if (*(save++) == 0) return false;
-	}
+    const wchar_t *save = s;
+    while (n-- > 0)
+    {
+        if (*(save++) == 0) return false;
+    }
 
-	return true;
+    return true;
 }
 
 // Custom types
@@ -97,22 +97,22 @@ typedef std::map<lite3dpp::WString, JSONValue*> JSONObject;
 
 class JSON
 {
-	friend class JSONValue;
-	
-	public:
-		static JSONValue* Parse(const char *data);
-		static JSONValue* Parse(const wchar_t *data);
-		static lite3dpp::WString Stringify(const JSONValue *value);
+    friend class JSONValue;
+    
+    public:
+        static JSONValue* Parse(const char *data);
+        static JSONValue* Parse(const wchar_t *data);
+        static lite3dpp::WString Stringify(const JSONValue *value);
         static lite3dpp::String wStringToString(const lite3dpp::WString &str);
         
-	protected:
-		static bool SkipWhitespace(const wchar_t **data);
+    protected:
+        static bool SkipWhitespace(const wchar_t **data);
         static bool SkipWhitespaceAndComments(const wchar_t **data);
-		static bool ExtractString(const wchar_t **data, lite3dpp::WString &str);
-		static double ParseInt(const wchar_t **data);
-		static double ParseDecimal(const wchar_t **data);
-	private:
-		JSON();
+        static bool ExtractString(const wchar_t **data, lite3dpp::WString &str);
+        static float ParseInt(const wchar_t **data);
+        static float ParseDecimal(const wchar_t **data);
+    private:
+        JSON();
 };
 
 #endif
